+++ src/pkg/services/m365/m365_test.go
@@ -1,17 +1,22 @@
-package m365_test
+package m365
 
 import (
+	"context"
 	"testing"
 
 	"github.com/alcionai/clues"
+	"github.com/microsoftgraph/msgraph-sdk-go/models"
+	"github.com/microsoftgraph/msgraph-sdk-go/models/odataerrors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 
+	"github.com/alcionai/corso/src/internal/common/ptr"
+	"github.com/alcionai/corso/src/internal/m365/discovery"
+	"github.com/alcionai/corso/src/internal/m365/graph"
 	"github.com/alcionai/corso/src/internal/tester"
 	"github.com/alcionai/corso/src/pkg/fault"
 	"github.com/alcionai/corso/src/pkg/path"
-	"github.com/alcionai/corso/src/pkg/services/m365"
 )
 
 type M365IntegrationSuite struct {
@@ -22,8 +27,7 @@ func TestM365IntegrationSuite(t *testing.T) {
 	suite.Run(t, &M365IntegrationSuite{
 		Suite: tester.NewIntegrationSuite(
 			t,
-			[][]string{tester.M365AcctCredEnvs},
-		),
+			[][]string{tester.M365AcctCredEnvs}),
 	})
 }
 
@@ -35,7 +39,7 @@ func (suite *M365IntegrationSuite) TestUsers() {
 
 	acct := tester.NewM365Account(suite.T())
 
-	users, err := m365.Users(ctx, acct, fault.New(true))
+	users, err := Users(ctx, acct, fault.New(true))
 	assert.NoError(t, err, clues.ToCore(err))
 	assert.NotEmpty(t, users)
 
@@ -59,7 +63,7 @@ func (suite *M365IntegrationSuite) TestUsersCompat_HasNoInfo() {
 
 	acct := tester.NewM365Account(suite.T())
 
-	users, err := m365.UsersCompatNoInfo(ctx, acct)
+	users, err := UsersCompatNoInfo(ctx, acct)
 	assert.NoError(t, err, clues.ToCore(err))
 	assert.NotEmpty(t, users)
 
@@ -85,7 +89,7 @@ func (suite *M365IntegrationSuite) TestGetUserInfo() {
 		uid  = tester.M365UserID(t)
 	)
 
-	info, err := m365.GetUserInfo(ctx, acct, uid)
+	info, err := GetUserInfo(ctx, acct, uid)
 	require.NoError(t, err, clues.ToCore(err))
 	require.NotNil(t, info)
 	require.NotEmpty(t, info)
@@ -112,7 +116,7 @@ func (suite *M365IntegrationSuite) TestUserHasMailbox() {
 		uid  = tester.M365UserID(t)
 	)
 
-	enabled, err := m365.UserHasMailbox(ctx, acct, uid)
+	enabled, err := UserHasMailbox(ctx, acct, uid)
 	require.NoError(t, err, clues.ToCore(err))
 	assert.True(t, enabled)
 }
@@ -128,7 +132,7 @@ func (suite *M365IntegrationSuite) TestUserHasDrive() {
 		uid  = tester.M365UserID(t)
 	)
 
-	enabled, err := m365.UserHasDrives(ctx, acct, uid)
+	enabled, err := UserHasDrives(ctx, acct, uid)
 	require.NoError(t, err, clues.ToCore(err))
 	assert.True(t, enabled)
 }
@@ -139,18 +143,219 @@ func (suite *M365IntegrationSuite) TestSites() {
 	ctx, flush := tester.NewContext(t)
 	defer flush()
 
-	acct := tester.NewM365Account(suite.T())
+	acct := tester.NewM365Account(t)
 
-	sites, err := m365.Sites(ctx, acct, fault.New(true))
+	sites, err := Sites(ctx, acct, fault.New(true))
 	assert.NoError(t, err, clues.ToCore(err))
 	assert.NotEmpty(t, sites)
 
 	for _, s := range sites {
-		suite.Run("site", func() {
+		suite.Run("site_"+s.ID, func() {
 			t := suite.T()
 			assert.NotEmpty(t, s.WebURL)
 			assert.NotEmpty(t, s.ID)
 			assert.NotEmpty(t, s.DisplayName)
+		})
+	}
+}
+
+type m365UnitSuite struct {
+	tester.Suite
+}
+
+func TestM365UnitSuite(t *testing.T) {
+	suite.Run(t, &m365UnitSuite{Suite: tester.NewUnitSuite(t)})
+}
+
+type mockDGDD struct {
+	response models.Driveable
+	err      error
+}
+
+func (m mockDGDD) GetDefaultDrive(context.Context, string) (models.Driveable, error) {
+	return m.response, m.err
+}
+
+func (suite *m365UnitSuite) TestCheckUserHasDrives() {
+	table := []struct {
+		name      string
+		mock      func(context.Context) discovery.GetDefaultDriver
+		expect    assert.BoolAssertionFunc
+		expectErr func(*testing.T, error)
+	}{
+		{
+			name: "ok",
+			mock: func(ctx context.Context) discovery.GetDefaultDriver {
+				return mockDGDD{models.NewDrive(), nil}
+			},
+			expect: assert.True,
+			expectErr: func(t *testing.T, err error) {
+				assert.NoError(t, err, clues.ToCore(err))
+			},
+		},
+		{
+			name: "mysite not found",
+			mock: func(ctx context.Context) discovery.GetDefaultDriver {
+				odErr := odataerrors.NewODataError()
+				merr := odataerrors.NewMainError()
+				merr.SetCode(ptr.To("code"))
+				merr.SetMessage(ptr.To(string(graph.MysiteNotFound)))
+				odErr.SetError(merr)
+
+				return mockDGDD{nil, graph.Stack(ctx, odErr)}
+			},
+			expect: assert.False,
+			expectErr: func(t *testing.T, err error) {
+				assert.NoError(t, err, clues.ToCore(err))
+			},
+		},
+		{
+			name: "mysite URL not found",
+			mock: func(ctx context.Context) discovery.GetDefaultDriver {
+				odErr := odataerrors.NewODataError()
+				merr := odataerrors.NewMainError()
+				merr.SetCode(ptr.To("code"))
+				merr.SetMessage(ptr.To(string(graph.MysiteURLNotFound)))
+				odErr.SetError(merr)
+
+				return mockDGDD{nil, graph.Stack(ctx, odErr)}
+			},
+			expect: assert.False,
+			expectErr: func(t *testing.T, err error) {
+				assert.NoError(t, err, clues.ToCore(err))
+			},
+		},
+		{
+			name: "no sharepoint license",
+			mock: func(ctx context.Context) discovery.GetDefaultDriver {
+				odErr := odataerrors.NewODataError()
+				merr := odataerrors.NewMainError()
+				merr.SetCode(ptr.To("code"))
+				merr.SetMessage(ptr.To(string(graph.NoSPLicense)))
+				odErr.SetError(merr)
+
+				return mockDGDD{nil, graph.Stack(ctx, odErr)}
+			},
+			expect: assert.False,
+			expectErr: func(t *testing.T, err error) {
+				assert.NoError(t, err, clues.ToCore(err))
+			},
+		},
+		{
+			name: "user not found",
+			mock: func(ctx context.Context) discovery.GetDefaultDriver {
+				odErr := odataerrors.NewODataError()
+				merr := odataerrors.NewMainError()
+				merr.SetCode(ptr.To(string(graph.RequestResourceNotFound)))
+				merr.SetMessage(ptr.To("message"))
+				odErr.SetError(merr)
+
+				return mockDGDD{nil, graph.Stack(ctx, odErr)}
+			},
+			expect: assert.False,
+			expectErr: func(t *testing.T, err error) {
+				assert.Error(t, err, clues.ToCore(err))
+			},
+		},
+		{
+			name: "arbitrary error",
+			mock: func(ctx context.Context) discovery.GetDefaultDriver {
+				odErr := odataerrors.NewODataError()
+				merr := odataerrors.NewMainError()
+				merr.SetCode(ptr.To("code"))
+				merr.SetMessage(ptr.To("message"))
+				odErr.SetError(merr)
+
+				return mockDGDD{nil, graph.Stack(ctx, odErr)}
+			},
+			expect: assert.False,
+			expectErr: func(t *testing.T, err error) {
+				assert.Error(t, err, clues.ToCore(err))
+			},
+		},
+	}
+	for _, test := range table {
+		suite.Run(test.name, func() {
+			t := suite.T()
+
+			ctx, flush := tester.NewContext(t)
+			defer flush()
+
+			dgdd := test.mock(ctx)
+
+			ok, err := checkUserHasDrives(ctx, dgdd, "foo")
+			test.expect(t, ok, "has drives flag")
+			test.expectErr(t, err)
+		})
+	}
+}
+
+type mockGAS struct {
+	response []models.Siteable
+	err      error
+}
+
+func (m mockGAS) GetAll(context.Context, *fault.Bus) ([]models.Siteable, error) {
+	return m.response, m.err
+}
+
+func (suite *m365UnitSuite) TestGetAllSites() {
+	table := []struct {
+		name      string
+		mock      func(context.Context) getAllSiteser
+		expectErr func(*testing.T, error)
+	}{
+		{
+			name: "ok",
+			mock: func(ctx context.Context) getAllSiteser {
+				return mockGAS{[]models.Siteable{}, nil}
+			},
+			expectErr: func(t *testing.T, err error) {
+				assert.NoError(t, err, clues.ToCore(err))
+			},
+		},
+		{
+			name: "no sharepoint license",
+			mock: func(ctx context.Context) getAllSiteser {
+				odErr := odataerrors.NewODataError()
+				merr := odataerrors.NewMainError()
+				merr.SetCode(ptr.To("code"))
+				merr.SetMessage(ptr.To(string(graph.NoSPLicense)))
+				odErr.SetError(merr)
+
+				return mockGAS{nil, graph.Stack(ctx, odErr)}
+			},
+			expectErr: func(t *testing.T, err error) {
+				assert.ErrorIs(t, err, graph.ErrServiceNotEnabled, clues.ToCore(err))
+			},
+		},
+		{
+			name: "arbitrary error",
+			mock: func(ctx context.Context) getAllSiteser {
+				odErr := odataerrors.NewODataError()
+				merr := odataerrors.NewMainError()
+				merr.SetCode(ptr.To("code"))
+				merr.SetMessage(ptr.To("message"))
+				odErr.SetError(merr)
+
+				return mockGAS{nil, graph.Stack(ctx, odErr)}
+			},
+			expectErr: func(t *testing.T, err error) {
+				assert.Error(t, err, clues.ToCore(err))
+			},
+		},
+	}
+	for _, test := range table {
+		suite.Run(test.name, func() {
+			t := suite.T()
+
+			ctx, flush := tester.NewContext(t)
+			defer flush()
+
+			gas := test.mock(ctx)
+
+			_, err := getAllSites(ctx, gas)
+			test.expectErr(t, err)
 		})
 	}
 }